Intron constructor includes the transcript
There are cases where two transcripts contain the same exons, which
causes the intron construction to fail.
e.g. The old code fails on the transcript ENSMUST00000116629

In this case, two exons do not uniquely define an intron, as an intron
is required to know which transcript it is part of. So we need to
specify the transcript.